backend/pyproject.toml
[tool.poetry]
name = "encurtador-link"
version = "0.1.0"
description = "Sistema de encurtamento de links"
authors = ["Lucas"]
readme = "README.md"
packages = [{include = "app.py"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
asyncpg = "^0.30.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
